requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
db-dtypes==1.2.0
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.0
flatbuffers==24.3.25
fonttools==4.51.0
functions-framework==3.5.0
fuzzysearch==0.7.3
gast==0.6.0
google-api-core==2.15.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-dialogflow-cx==1.21.0
google-cloud-discoveryengine==0.9.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kaleido==0.2.1
keras==2.15.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==23.2
pandas==2.2.2
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.22.0
plotly-express==0.4.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
pypdf==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.13.0
Shapely==1.8.5.post1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
tenacity==8.3.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
Unidecode==1.3.8
urllib3==2.1.0
uvicorn==0.30.1
uvloop==0.19.0
watchdog==3.0.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
